+++ packages/google_sign_in/google_sign_in_platform_interface/lib/google_sign_in_platform_interface.dart
@@ -86,14 +86,7 @@ abstract class GoogleSignInPlatform {
     String? hostedDomain,
     String? clientId,
   }) async {
-    await initWithParams(
-      SignInInitParameters(
-        scopes: scopes,
-        signInOption: signInOption,
-        hostedDomain: hostedDomain,
-        clientId: clientId,
-      ),
-    );
+    throw UnimplementedError('init() has not been implemented.');
   }
 
   /// Initializes the plugin. You must call this method before calling other
@@ -106,7 +99,12 @@ abstract class GoogleSignInPlatform {
   ///
   /// * [SignInInitParameters]
   Future<void> initWithParams(SignInInitParameters params) async {
-    throw UnimplementedError('initWithParams() has not been implemented');
+    await init(
+      scopes: params.scopes,
+      signInOption: params.signInOption,
+      hostedDomain: params.hostedDomain,
+      clientId: params.clientId,
+    );
   }
 
   /// Attempts to reuse pre-existing credentials to sign in again, without user interaction.