+++ packages/google_sign_in/google_sign_in_platform_interface/test/method_channel_google_sign_in_test.dart
@@ -139,7 +139,7 @@ void main() {
     });
 
     test('initWithParams passes through arguments to the channel', () async {
-      await googleSignIn.initWithParams(SignInInitParameters(
+      await googleSignIn.initWithParams(const SignInInitParameters(
           hostedDomain: 'example.com',
           scopes: <String>['two', 'scopes'],
           signInOption: SignInOption.games,