+++ packages/google_sign_in/google_sign_in_platform_interface/CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 2.1.3
 
 * Removes unnecessary imports.
-* Add `SignInInitParameters` class to hold all sign in params, including the new `forceCodeForRefreshToken`.
+* Adds `SignInInitParameters` class to hold all sign in params, including the new `forceCodeForRefreshToken`.
 
 ## 2.1.2